+++ R/reviewPackage.R
@@ -18,6 +18,8 @@ globalVariables(
     "n",
     "commit",
     "name",
-    "value"
+    "value",
+    "paths",
+    "time"
   )
 )