+++ tests/testthat/test-logSummary.R
@@ -7,5 +7,10 @@ with_demoRepo({
     expect_true(tempdf_sum %>% dplyr::filter(file == "script/data-assembly.R") %>% dplyr::pull(headf) == 5)
     expect_true(nrow(tempdf_sum %>% dplyr::count(file) %>% dplyr::count(n, name = "num_files")) == 1)
   })
+  
+  test_that("logSummary prints as a data.frame", {
+    expect_true(inherits(logSummary(), "data.frame"))
+  })
+  
 })