+++ NEWS.md
@@ -2,7 +2,9 @@
 
 ## New features and changes
 
-- `renderQCSummary` summary tables updated. 
+- `diffFiles` can now display the entire file that is being diffed. (#115)
+
+- Added `repoHistory` function to return history of all commits in the repository. (#116)
 
 # review 3.9.1